database/requirements.txt
# requirements.txt
psycopg2-binary
meteofrance_api